5. Fraude Detection/IEEE-CIS Fraud Detection - InProgress/requirements.txt
bleach==6.3.0
catboost==1.2.10
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fonttools==4.62.1
graphviz==0.21
idna==3.11
joblib==1.5.3
kaggle==2.0.0
kagglehub==1.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
lightgbm==4.6.0
matplotlib==3.10.8
narwhals==2.18.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pipreqs==0.4.13
plotly==6.6.0
protobuf==7.34.1
pyarrow==23.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.3
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
xgboost==3.2.0
yarg==0.1.10
